refactor demo.py
I will be adding an interactive demo
and it makes sense to create a demo
module
